refactor: streamline database model definitions and imports
- Removed unused imports and consolidated SQLAlchemy model definitions for improved clarity and maintainability.
- Simplified the import structure by eliminating unnecessary type checking and redundant model declarations.
- Enhanced the organization of model imports to facilitate easier access and understanding of the database schema.
- These changes contribute to a cleaner codebase and better performance in database interactions.
